Allow additional volumes on the operator
Allowing additional volumes on the operator means we can attach more
certificates without having to build them into the image. Once we can
get the certificates into an operator container we can move onto
configuring the redis clients used in the operator to use the
certificates when configured to do so.